GRAL: Se crearon las validaciones para el plan

+++ database/migrations/2020_05_22_212333_create_people_table.php
@@ -22,7 +22,7 @@ class CreatePeopleTable extends Migration
             $table->integer('age')->nullable();;
             $table->string('email',30);
             $table->integer('weight');
-            $table->integer('height');
+            $table->float('height');
             $table->timestamps();
         });
     }
